+++ CHANGES.rst
@@ -31,14 +31,20 @@ Unreleased
   (arbitrary sequence of marked lines).  See the :ref:`multi_line_exclude`
   section of the docs for more details and examples.
 
+- The JSON report now includes per-function and per-class coverage information.
+  Thanks to `Daniel Diniz <pull 1809_>`_ for getting the work started.
+
+- Fixed an incorrect calculation of "(no class)" lines in the HTML classes
+  report.
+
 - Python 3.13.0b3 is supported.
 
 .. _issue 118: https://github.com/nedbat/coveragepy/issues/118
 .. _issue 996: https://github.com/nedbat/coveragepy/issues/996
 .. _issue 1741: https://github.com/nedbat/coveragepy/issues/1741
 .. _issue 1803: https://github.com/nedbat/coveragepy/issues/1803
-
 .. _pull 1807: https://github.com/nedbat/coveragepy/pull/1807
+.. _pull 1809: https://github.com/nedbat/coveragepy/pull/1809
 
 .. scriv-start-here